Fix dabble home (#15)
* Get user.home from dsl

Use the DSL to get the `user.home` system property. Previously this was
a direct side-effect in the code.

* Update dabble home

Fixed Dabble's home directory to poin to `~/.dabble`. In the previous
release Dabble's home directory was incorrectly pointing to `~`.

Cleaned up Path type names and pull out path Strings into constants.

* Bump version to 0.4.1